5 dec - Finished
Part 1 was done quick, the same I thought of Part 2. Until it kept running and didn't finish. I quickly noticed the memes on Reddit and it probably would've takes hours if not days/months to solve and petabytes of storage. Came up with a quicker algorithm, which took a while to refine as I missed an edge-case. Got it now though!